image_embeddings_pipeline/requirements.txt
python-dotenv
pandas 
requests 
pillow 
qdrant-client
aiohttp